Move carousel component to cite folder
