+++ site/gatsby-site/src/components/cite/ImageCarousel.js
