+++ site/gatsby-site/src/templates/cite.js
@@ -4,7 +4,7 @@ import { Button, Col, Container, Row } from 'react-bootstrap';
 import Layout from 'components/Layout';
 import { StyledHeading } from 'components/styles/Docs';
 import Citation from 'components/cite/Citation';
-import ImageCarousel from 'components/ImageCarousel';
+import ImageCarousel from 'components/cite/ImageCarousel';
 import BibTex from 'components/BibTex';
 import { getCanonicalUrl } from 'utils/getCanonicalUrl';
 import styled from 'styled-components';
